outputs where generated, srs was changed

+++ testing/test_user.cpp
@@ -6,7 +6,7 @@ TEST_CASE( "Test Users", "[User]" ) {
 	Job * job = new Job(config, now, SIXTY_FOUR_HOURS);
 
 	User * support = new User(config, 1, true);
-	for (int i = 0; i < 20; i++) {
+	for (size_t i = 0; i < 20; i++) {
 		support->pay(job);
 	}
 	REQUIRE(support->can_afford(job));
